backend/requirements.txt
Django==3.2.9
django-chunked-upload
python-magic
django-cors-headers
requests
beautifulsoup4
